remove-dangling-images: Check if exists before removing file

+++ scripts/remove-dangling-images.xsh
@@ -31,5 +31,6 @@ try:
 finally:
     fcntl.flock(lock_fd, fcntl.LOCK_UN)
     lock_fd.close()
-    os.remove(lock_path)
+    if os.path.exists(lock_path):
+        os.remove(lock_path)
